gather lib/tangles
