+++ lib/tangles/add-tangles.js
@@ -5,7 +5,7 @@
 const set = require('lodash.set')
 const clarify = require('clarify-error')
 const { isIdentityGroupSSBURI } = require('ssb-uri2')
-const getTangle = require('./get-tangle')
+const getTangleData = require('./get-tangle-data')
 
 /**
  * Note that this mutates `content`
@@ -17,7 +17,7 @@ module.exports = function addTangles(server, content, tangles, cb) {
 
     const currTangle = tangles[0]
 
-    getTangle(server, currTangle, content.recps[0], (err, generatedTangle) => {
+    getTangleData(server, currTangle, content.recps[0], (err, generatedTangle) => {
       // prettier-ignore
       if (err) return cb(clarify(err, 'Failed to get group tangle when adding group tangle to message'))
