+++ lib/tangles/get-tangle-data.js
@@ -2,12 +2,14 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-only
 
-const pull = require('pull-stream')
 const Reduce = require('@tangle/reduce')
 const Strategy = require('@tangle/strategy')
 const clarify = require('clarify-error')
 const { isIdentityGroupSSBURI, fromMessageSigil } = require('ssb-uri2')
 const { where, author, toPullStream } = require('ssb-db2/operators')
+const pull = require('pull-stream')
+
+const getUpdates = require('./get-tangle-updates')
 
 const strategy = new Strategy({})
 
@@ -62,7 +64,7 @@ function getTangleRoot(server, groupId, tangle, cb) {
 }
 
 /** for figuring out what "previous" should be for the group. `server` is the ssb server you're using. `tangle` is the name of the tangle in the group you're looking for, e.g. "group" or "members" */
-function getTangle(server, tangle, groupId, cb) {
+module.exports = function getTangleData(server, tangle, groupId, cb) {
   if (!isIdentityGroupSSBURI(groupId)) {
     // prettier-ignore
     return cb(new Error(`get-tangle expects valid groupId, got: ${groupId}`))
@@ -96,39 +98,4 @@ function getTangle(server, tangle, groupId, cb) {
       })
     })
   })
-}
-module.exports = getTangle
-
-const B_VALUE = Buffer.from('value')
-const B_CONTENT = Buffer.from('content')
-const B_TANGLES = Buffer.from('tangles')
-const B_ROOT = Buffer.from('root')
-
-function getUpdates(ssb, tangle, root, cb) {
-  const { seekKey } = require('bipf')
-  const B_TANGLE = Buffer.from(tangle)
-
-  function tangleRoot(value) {
-    return ssb.db.operators.equal(seekTangleRoot, value, {
-      indexType: `value.content.tangles.${tangle}.root`,
-    })
-  }
-  function seekTangleRoot(buffer) {
-    let p = 0 // note you pass in p!
-    p = seekKey(buffer, p, B_VALUE)
-    if (p < 0) return
-    p = seekKey(buffer, p, B_CONTENT)
-    if (p < 0) return
-    p = seekKey(buffer, p, B_TANGLES)
-    if (p < 0) return
-    p = seekKey(buffer, p, B_TANGLE)
-    if (p < 0) return
-    return seekKey(buffer, p, B_ROOT)
-  }
-
-  const { where, and, toPullStream } = ssb.db.operators
-  pull(
-    ssb.db.query(where(and(tangleRoot(root))), toPullStream()),
-    pull.collect(cb)
-  )
 }
